msm: mdss: Use 64-bit math for bandwidth calculation
For larger panel with longer vertical pixel count, it is possible that
bandwidth requirement is much higher during downscaling usecase. Current
bandwidth calculation for compression ratio adjustment only use 32-bit
and it is shown that overflow can happen, and is required to change to
use 64-bit variable as input and output.

CRs-Fixed: 2045602
Change-Id: I817e9d55fb6e24e686513327d00f7efd08ac717f
Signed-off-by: Benjamin Chan <bkchan@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>